// src/app/open-api/generated/model/document-column.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DocumentColumn { 
    readonly id?: number;
    name: string;
    displayedName: string;
    dataType?: DocumentColumn.DataTypeEnum;
    order: number;
    visibility: boolean;
    editable: boolean;
}
export namespace DocumentColumn {
    export const DataTypeEnum = {
        Date: 'DATE',
        String: 'STRING',
        Number: 'NUMBER',
        Object: 'OBJECT'
    } as const;
    export type DataTypeEnum = typeof DataTypeEnum[keyof typeof DataTypeEnum];
}